removed colorNumChannel from any setup functions. now it is in set(), so you can dynamically swap from color to bw while running. the number of channels is also irrelevant to everything but the set() function.

+++ trunk/OpenFrameworks/addons/ofxStructuredLight/src/ThreePhaseGenerator.h
@@ -7,7 +7,7 @@ class ThreePhaseGenerator : public PatternGenerator {
 private:
 	float wavelength;
 	phaseOrientation orientation;
-	int minBrightness;
+	float minBrightness;
 public:
 
 	void setMinBrightness(float minBrightness);
